requirements.txt
torch==2.3.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==2.1.1
click==8.2.1
FastAPI
uvicorn[standard]
uvloop
httpx
requests
Flask==2.3.2
gunicorn==23.0.0
greenlet==3.2.4
aiomysql
tensorflow-cpu==2.12.0
pymysql
python-dotenv
cryptography
idna==3.10
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
packaging==25.0
SQLAlchemy==1.4.53
typing_extensions==4.8.0
urllib3==1.26.20
lime
onnxruntime
onnx
numpy
